14. Ukrycie klasy Node w IntLinkedList
