+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -42,4 +42,30 @@ public class IntLinkedList {
         }
         return ret;
     }
+
+    class Node {
+
+        private final int value;
+        private Node prev, next;
+
+        public Node(int value) {
+            this.value = value;
+        }
+
+        public int getValue() {
+            return value;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
+    }
 }
